chore(story): :zap: update file to tsx, add another option using Ref
